added ValidationBlockInstruction to allow blocks to opt-in validation
