+++ packages/schema-blocks/src/core/blocks/BlockInstructionClass.ts
@@ -1,0 +1,7 @@
+import { InstructionOptions } from "../Instruction";
+import BlockInstruction from "./BlockInstruction";
+
+export type BlockInstructionClass = {
+	new( id: number, options: InstructionOptions ): BlockInstruction;
+	options: InstructionOptions;
+};
